Big refactor to get the code into the world of modern javascript
