+++ README.md
@@ -2,3 +2,11 @@
 =======
 
 Cycle-accurate 6502 emulator in Javascript
+
+
+BUILDING
+=======
+
+Run `npm install` first to retrieve all dependencies.
+Running `npm start` builds both the minified and the development version of 6502.js.
+`npm run build:production` and `npm run build:production` build the minified and development version separately.
