+++ src/opcodes.js
@@ -1,641 +1,3 @@
-CPU6502 = function() {
-
-	this.PC = 0; // Program counter
-
-	this.A = 0; this.X = 0; this.Y = 0; this.S = 0; // Registers
-	this.N = 0; this.Z = 1; this.C = 0; this.V = 0; // ALU flags
-	this.I = 0; this.D = 0; // Other flags
-
-	this.irq = 0; this.nmi = 0; // IRQ lines
-
-	this.tmp = 0; this.addr = 0; // Temporary registers
-	this.opcode = 0; // Current opcode
-	this.cycles = 0; // Cycles counter
-}
-
-CPU6502.constructor = CPU6502;
-
-////////////////////////////////////////////////////////////////////////////////
-// CPU control
-////////////////////////////////////////////////////////////////////////////////
-
-CPU6502.prototype.reset = function() {
-
-	this.A = 0; this.X = 0; this.Y = 0; this.S = 0;
-	this.N = 0; this.Z = 1; this.C = 0; this.V = 0;
-	this.I = 0; this.D = 0;
-
-	this.PC = (this.read(0xFFFD) << 8) | this.read(0xFFFC);
-}
-
-CPU6502.prototype.step = function() {
-	this.opcode = this.read( this.PC++ );
-	CPU6502op[ this.opcode ]( this );
-}
-
-CPU6502.prototype.log = function() {
-	var msg = "nPC=" + this.PC.toString(16);
-	msg += " cyc=" + this.cycles;
-	msg += " [" + this.opcode.toString(16) + "] ";
-	msg += ( this.C ? "C" : "-");
-	msg += ( this.N ? "N" : "-");
-	msg += ( this.Z ? "Z" : "-");
-	msg += ( this.V ? "V" : "-");
-	msg += ( this.D ? "D" : "-");
-	msg += ( this.I ? "I" : "-");
-	msg += " A=" + this.A.toString(16);
-	msg += " X=" + this.X.toString(16);
-	msg += " Y=" + this.Y.toString(16);
-	msg += " S=" + this.S.toString(16);
-	console.log(msg);
-}
-
-////////////////////////////////////////////////////////////////////////////////
-// Subroutines - addressing modes & flags
-////////////////////////////////////////////////////////////////////////////////
-
-CPU6502.prototype.izx = function() {
-	var a = (this.read(this.PC++) + this.X) & 0xFF;
-	this.addr = (this.read(a+1) << 8) | this.read(a);
-	this.cycles += 6;
-}
-
-CPU6502.prototype.izy = function() {
-	var a = this.read(this.PC++);
-	var paddr = (this.read((a+1) & 0xFF) << 8) | this.read(a);
-	this.addr = (paddr + this.Y);
-	if ( (paddr & 0x100) != (this.addr & 0x100) ) {
-		this.cycles += 6;
-	} else {
-		this.cycles += 5;
-	}
-}
-
-CPU6502.prototype.ind = function() {
-	var a = this.read(this.PC);
-	a |= this.read( (this.PC & 0xFF00) | ((this.PC + 1) & 0xFF) ) << 8;
-	this.addr = this.read(a);
-	this.addr |= (this.read(a+1) << 8);
-	this.cycles += 5;
-}
-
-CPU6502.prototype.zp = function() {
-	this.addr = this.read(this.PC++);
-	this.cycles += 3;
-}
-
-CPU6502.prototype.zpx = function() {
-	this.addr = (this.read(this.PC++) + this.X) & 0xFF;
-	this.cycles += 4;
-}
-
-CPU6502.prototype.zpy = function() {
-	this.addr = (this.read(this.PC++) + this.Y) & 0xFF;
-	this.cycles += 4;
-}
-
-CPU6502.prototype.imp = function() {
-	this.cycles += 2;
-}
-
-CPU6502.prototype.imm = function() {
-	this.addr = this.PC++;
-	this.cycles += 2;
-}
-
-CPU6502.prototype.abs = function() {
-	this.addr = this.read(this.PC++);
-	this.addr |= (this.read(this.PC++) << 8);
-	this.cycles += 4;
-}
-
-CPU6502.prototype.abx = function() {
-	var paddr = this.read(this.PC++);
-	paddr |= (this.read(this.PC++) << 8);
-	this.addr = (paddr + this.X);
-	if ( (paddr & 0x100) != (this.addr & 0x100) ) {
-		this.cycles += 5;
-	} else {
-		this.cycles += 4;
-	}
-}
-
-CPU6502.prototype.aby = function() {
-	var paddr = this.read(this.PC++);
-	paddr |= (this.read(this.PC++) << 8);
-	this.addr = (paddr + this.Y);
-	if ( (paddr & 0x100) != (this.addr & 0x100) ) {
-		this.cycles += 5;
-	} else {
-		this.cycles += 4;
-	}
-}
-
-CPU6502.prototype.rel = function() {
-	this.addr = this.read(this.PC++);
-	if (this.addr & 0x80) {
-		this.addr -= 0x100;
-	}
-	this.addr += this.PC;
-	this.cycles += 2;
-}
-
-////////////////////////////////////////////////////////////////////////////////
-
-CPU6502.prototype.rmw = function() {
-	this.write(this.addr, this.tmp & 0xFF);
-	this.cycles += 2;
-}
-
-////////////////////////////////////////////////////////////////////////////////
-
-CPU6502.prototype.fnz = function(v) {
-	this.Z = ((v & 0xFF) == 0) ? 1 : 0;
-	this.N = ((v & 0x80) != 0) ? 1 : 0;
-}
-
-// Borrow
-CPU6502.prototype.fnzb = function(v) {
-	this.Z = ((v & 0xFF) == 0) ? 1 : 0;
-	this.N = ((v & 0x80) != 0) ? 1 : 0;
-	this.C = ((v & 0x100) != 0) ? 0 : 1;
-}
-
-// Carry
-CPU6502.prototype.fnzc = function(v) {
-	this.Z = ((v & 0xFF) == 0) ? 1 : 0;
-	this.N = ((v & 0x80) != 0) ? 1 : 0;
-	this.C = ((v & 0x100) != 0) ? 1 : 0;
-}
-
-CPU6502.prototype.branch = function(v) {
-	if (v) {
-		if ( (this.addr & 0x100) != (this.PC & 0x100) ) {
-			this.cycles += 2;
-		} else {
-			this.cycles += 1;
-		}
-		this.PC = this.addr;
-	}
-}
-
-////////////////////////////////////////////////////////////////////////////////
-// Subroutines - instructions
-////////////////////////////////////////////////////////////////////////////////
-CPU6502.prototype.adc = function() {
-	var v = this.read(this.addr);
-	var c = this.C;
-	var r = this.A + v + c;
-	if (this.D) {
-		var al = (this.A & 0x0F) + (v & 0x0F) + c;
-		if (al > 9) al += 6;
-		var ah = (this.A >> 4) + (v >> 4) + ((al > 15) ? 1 : 0);
-		this.Z = ((r & 0xFF) == 0) ? 1 : 0;
-		this.N = ((ah & 8) != 0) ? 1 : 0;
-		this.V = ((~(this.A ^ v) & (this.A ^ (ah << 4)) & 0x80) != 0) ? 1 : 0;
-		if (ah > 9) ah += 6;
-		this.C = (ah > 15) ? 1 : 0;
-		this.A = ((ah << 4) | (al & 15)) & 0xFF;
-	} else {
-		this.Z = ((r & 0xFF) == 0) ? 1 : 0;
-		this.N = ((r & 0x80) != 0) ? 1 : 0;
-		this.V = ((~(this.A ^ v) & (this.A ^ r) & 0x80) != 0) ? 1 : 0;
-		this.C = ((r & 0x100) != 0) ? 1 : 0;
-		this.A = r & 0xFF;
-	}
-}
-
-CPU6502.prototype.ahx = function() {
-	this.tmp = ((this.addr >> 8) + 1) & this.A & this.X;
-	this.write(this.addr, this.tmp & 0xFF);
-}
-
-CPU6502.prototype.alr = function() {
-	this.tmp = this.read(this.addr) & this.A;
-	this.tmp = ((this.tmp & 1) << 8) | (this.tmp >> 1);
-	this.fnzc(this.tmp);
-	this.A = this.tmp & 0xFF;
-}
-
-CPU6502.prototype.anc = function() {
-	this.tmp = this.read(this.addr);
-	this.tmp |= ((this.tmp & 0x80) & (this.A & 0x80)) << 1;
-	this.fnzc(this.tmp);
-	this.A = this.tmp & 0xFF;
-}
-
-CPU6502.prototype.and = function() {
-	this.A &= this.read(this.addr);
-	this.fnz(this.A);
-}
-
-CPU6502.prototype.ane = function() {
-	this.tmp = this.read(this.addr) & this.A & (this.A | 0xEE);
-	this.fnz(this.tmp);
-	this.A = this.tmp & 0xFF;
-}
-
-CPU6502.prototype.arr = function() {
-	this.tmp = this.read(this.adfdr) & this.A;
-	this.C = ((this.tmp & 0x80) != 0);
-	this.V = ((((this.tmp >> 7) & 1) ^ ((this.tmp >> 6) & 1)) != 0);
-	if (this.D) {
-		var al = (this.tmp & 0x0F) + (this.tmp & 1);
-		if (al > 5) al += 6;
-		var ah = ((this.tmp >> 4) & 0x0F) + ((this.tmp >> 4) & 1);
-		if (ah > 5) {
-			al += 6;
-			this.C = true;
-		} else {
-			this.C = false;
-		}
-		this.tmp = (ah << 4) | al;
-	}
-	this.fnz(this.tmp);
-	this.A = this.tmp & 0xFF;
-}
-
-CPU6502.prototype.asl = function() {
-	this.tmp = this.read(this.addr) << 1;
-	this.fnzc(this.tmp);
-	this.tmp &= 0xFF;
-}
-CPU6502.prototype.asla = function() {
-	this.tmp = this.A << 1;
-	this.fnzc(this.tmp);
-	this.A = this.tmp & 0xFF;
-}
-
-CPU6502.prototype.bit = function() {
-	this.tmp = this.read(this.addr);
-	this.N = ((this.tmp & 0x80) != 0) ? 1 : 0;
-	this.V = ((this.tmp & 0x40) != 0) ? 1 : 0;
-	this.Z = ((this.tmp & this.A) == 0) ? 1 : 0;
-}
-
-CPU6502.prototype.brk = function() {
-	this.PC++;
-	this.write(this.S + 0x100, this.PC >> 8);
-	this.S = (this.S - 1) & 0xFF;
-	this.write(this.S + 0x100, this.PC & 0xFF);
-	this.S = (this.S - 1) & 0xFF;
-	var v = this.N << 7;
-	v |= this.V << 6;
-	v |= 3 << 4;
-	v |= this.D << 3;
-	v |= this.I << 2;
-	v |= this.Z << 1;
-	v |= this.C;
-	this.write(this.S + 0x100, v);
-	this.S = (this.S - 1) & 0xFF;
-	this.I = 1;
-	this.D = 0;
-	this.PC = (this.read(0xFFFF) << 8) | this.read(0xFFFE);
-	this.cycles += 5;
-}
-
-CPU6502.prototype.bcc = function() { this.branch( this.C == 0 ); }
-CPU6502.prototype.bcs = function() { this.branch( this.C == 1 ); }
-CPU6502.prototype.beq = function() { this.branch( this.Z == 1 ); }
-CPU6502.prototype.bne = function() { this.branch( this.Z == 0 ); }
-CPU6502.prototype.bmi = function() { this.branch( this.N == 1 ); }
-CPU6502.prototype.bpl = function() { this.branch( this.N == 0 ); }
-CPU6502.prototype.bvc = function() { this.branch( this.V == 0 ); }
-CPU6502.prototype.bvs = function() { this.branch( this.V == 1 ); }
-
-
-CPU6502.prototype.clc = function() { this.C = 0; }
-CPU6502.prototype.cld = function() { this.D = 0; }
-CPU6502.prototype.cli = function() { this.I = 0; }
-CPU6502.prototype.clv = function() { this.V = 0; }
-
-CPU6502.prototype.cmp = function() {
-	this.fnzb( this.A - this.read(this.addr) );
-}
-
-CPU6502.prototype.cpx = function() {
-	this.fnzb( this.X - this.read(this.addr) );
-}
-
-CPU6502.prototype.cpy = function() {
-	this.fnzb( this.Y - this.read(this.addr) );
-}
-
-CPU6502.prototype.dcp = function() {
-	this.tmp = (this.read(this.addr) - 1) & 0xFF;
-	this.tmp = this.A - this.tmp;
-	this.fnz(this.tmp);
-}
-
-CPU6502.prototype.dec = function() {
-	this.tmp = (this.read(this.addr) - 1) & 0xFF;
-	this.fnz(this.tmp);
-}
-
-CPU6502.prototype.dex = function() {
-	this.X = (this.X - 1) & 0xFF;
-	this.fnz(this.X);
-}
-
-CPU6502.prototype.dey = function() {
-	this.Y = (this.Y - 1) & 0xFF;
-	this.fnz(this.Y);
-}
-
-CPU6502.prototype.eor = function() {
-	this.A ^= this.read(this.addr);
-	this.fnz(this.A);
-}
-
-CPU6502.prototype.inc = function() {
-	this.tmp = (this.read(this.addr) + 1) & 0xFF;
-	this.fnz(this.tmp);
-}
-
-CPU6502.prototype.inx = function() {
-	this.X = (this.X + 1) & 0xFF;
-	this.fnz(this.X);
-}
-
-CPU6502.prototype.iny = function() {
-	this.Y = (this.Y + 1) & 0xFF;
-	this.fnz(this.Y);
-}
-
-CPU6502.prototype.isc = function() {
-	var v = (this.read(this.addr) + 1) & 0xFF;
-	var c = 1 - (this.C ? 1 : 0);
-	var r = this.A - v - c;
-	if (this.D) {
-		var al = (this.A & 0x0F) - (v & 0x0F) - c;
-		if (al > 0x80) al -= 6;
-		var ah = (this.A >> 4) - (v >> 4) - ((al > 0x80) ? 1 : 0);
-		this.Z = ((r & 0xFF) == 0);
-		this.N = ((r & 0x80) != 0);
-		this.V = (((this.A ^ v) & (this.A ^ r) & 0x80) != 0);
-		this.C = ((this.r & 0x100) != 0) ? 0 : 1;
-		if (ah > 0x80) ah -= 6;
-		this.A = ((ah << 4) | (al & 15)) & 0xFF;
-	} else {
-		this.Z = ((r & 0xFF) == 0);
-		this.N = ((r & 0x80) != 0);
-		this.V = (((this.A ^ v) & (this.A ^ r) & 0x80) != 0);
-		this.C = ((r & 0x100) != 0) ? 0 : 1;
-		this.A = r & 0xFF;
-	}
-}
-
-CPU6502.prototype.jmp = function() {
-	this.PC = this.addr;
-	this.cycles--;
-}
-
-CPU6502.prototype.jsr = function() {
-	this.write(this.S + 0x100, (this.PC - 1) >> 8);
-	this.S = (this.S - 1) & 0xFF;
-	this.write(this.S + 0x100, (this.PC - 1) & 0xFF);
-	this.S = (this.S - 1) & 0xFF;
-	this.PC = this.addr;
-	this.cycles += 2;
-}
-
-CPU6502.prototype.las = function() {
-	this.S = this.X = this.A = this.read(this.addr) & this.S;
-	this.fnz(this.A);
-}
-
-CPU6502.prototype.lax = function() {
-	this.X = this.A = this.read(this.addr);
-	this.fnz(this.A);
-}
-
-CPU6502.prototype.lda = function() {
-	this.A = this.read(this.addr);
-	this.fnz(this.A);
-}
-
-CPU6502.prototype.ldx = function() {
-	this.X = this.read(this.addr);
-	this.fnz(this.X);
-}
-
-CPU6502.prototype.ldy = function() {
-	this.Y = this.read(this.addr);
-	this.fnz(this.Y);
-}
-
-CPU6502.prototype.ora = function() {
-	this.A |= this.read(this.addr);
-	this.fnz(this.A);
-}
-
-CPU6502.prototype.rol = function() {
-	this.tmp = (this.read(this.addr) << 1) | this.C;
-	this.fnzc(this.tmp);
-	this.tmp &= 0xFF;
-}
-CPU6502.prototype.rla = function() {
-	this.tmp = (this.A << 1) | this.C;
-	this.fnzc(this.tmp);
-	this.A = this.tmp & 0xFF;
-}
-
-CPU6502.prototype.ror = function() {
-	this.tmp = this.read(this.addr);
-	this.tmp = ((this.tmp & 1) << 8) | (this.C << 7) | (this.tmp >> 1);
-	this.fnzc(this.tmp);
-	this.tmp &= 0xFF;
-}
-CPU6502.prototype.rra = function() {
-	this.tmp = ((this.A & 1) << 8) | (this.C << 7) | (this.A >> 1);
-	this.fnzc(this.tmp);
-	this.A = this.tmp & 0xFF;
-}
-
-
-CPU6502.prototype.lsr = function() {
-	this.tmp = this.read(this.addr);
-	this.tmp = ((this.tmp & 1) << 8) | (this.tmp >> 1);
-	this.fnzc(this.tmp);
-	this.tmp &= 0xFF;
-}
-CPU6502.prototype.lsra = function() {
-	this.tmp = ((this.A & 1) << 8) | (this.A >> 1);
-	this.fnzc(this.tmp);
-	this.A = this.tmp & 0xFF;
-}
-
-
-CPU6502.prototype.nop = function() { }
-
-CPU6502.prototype.pha = function() {
-	this.write(this.S + 0x100, this.A);
-	this.S = (this.S - 1) & 0xFF;
-	this.cycles++;
-}
-
-CPU6502.prototype.php = function() {
-	var v = this.N << 7;
-	v |= this.V << 6;
-	v |= 3 << 4;
-	v |= this.D << 3;
-	v |= this.I << 2;
-	v |= this.Z << 1;
-	v |= this.C;
-	this.write(this.S + 0x100, v);
-	this.S = (this.S - 1) & 0xFF;
-	this.cycles++;
-}
-
-CPU6502.prototype.pla = function() {
-	this.S = (this.S + 1) & 0xFF;
-	this.A = this.read(this.S + 0x100);
-	this.fnz(this.A);
-	this.cycles += 2;
-}
-
-CPU6502.prototype.plp = function() {
-	this.S = (this.S + 1) & 0xFF;
-	this.tmp = this.read(this.S + 0x100);
-	this.N = ((this.tmp & 0x80) != 0) ? 1 : 0;
-	this.V = ((this.tmp & 0x40) != 0) ? 1 : 0;
-	this.D = ((this.tmp & 0x08) != 0) ? 1 : 0;
-	this.I = ((this.tmp & 0x04) != 0) ? 1 : 0;
-	this.Z = ((this.tmp & 0x02) != 0) ? 1 : 0;
-	this.C = ((this.tmp & 0x01) != 0) ? 1 : 0;
-	this.cycles += 2;
-}
-
-CPU6502.prototype.rti = function() {
-	this.S = (this.S + 1) & 0xFF;
-	this.tmp = this.read(this.S + 0x100);
-	this.N = ((this.tmp & 0x80) != 0) ? 1 : 0;
-	this.V = ((this.tmp & 0x40) != 0) ? 1 : 0;
-	this.D = ((this.tmp & 0x08) != 0) ? 1 : 0;
-	this.I = ((this.tmp & 0x04) != 0) ? 1 : 0;
-	this.Z = ((this.tmp & 0x02) != 0) ? 1 : 0;
-	this.C = ((this.tmp & 0x01) != 0) ? 1 : 0;
-	this.S = (this.S + 1) & 0xFF;
-	this.PC = this.read(this.S + 0x100);
-	this.S = (this.S + 1) & 0xFF;
-	this.PC |= this.read(this.S + 0x100) << 8;
-	this.cycles += 4;
-}
-
-CPU6502.prototype.rts = function() {
-	this.S = (this.S + 1) & 0xFF;
-	this.PC = this.read(this.S + 0x100);
-	this.S = (this.S + 1) & 0xFF;
-	this.PC |= this.read(this.S + 0x100) << 8;
-	this.PC++;
-	this.cycles += 4;
-}
-
-CPU6502.prototype.sbc = function() {
-	var v = this.read(this.addr);
-	var c = 1 - this.C;
-	var r = this.A - v - c;
-	if (this.D) {
-		var al = (this.A & 0x0F) - (v & 0x0F) - c;
-		if (al < 0) al -= 6;
-		var ah = (this.A >> 4) - (v >> 4) - ((al < 0) ? 1 : 0);
-		this.Z = ((r & 0xFF) == 0) ? 1 : 0;
-		this.N = ((r & 0x80) != 0) ? 1 : 0;
-		this.V = (((this.A ^ v) & (this.A ^ r) & 0x80) != 0) ? 1 : 0;
-		this.C = ((r & 0x100) != 0) ? 0 : 1;
-		if (ah < 0) ah -= 6;
-		this.A = ((ah << 4) | (al & 15)) & 0xFF;
-	} else {
-		this.Z = ((r & 0xFF) == 0) ? 1 : 0;
-		this.N = ((r & 0x80) != 0) ? 1 : 0;
-		this.V = (((this.A ^ v) & (this.A ^ r) & 0x80) != 0) ? 1 : 0;
-		this.C = ((r & 0x100) != 0) ? 0 : 1;
-		this.A = r & 0xFF;
-	}
-}
-
-CPU6502.prototype.sbx = function() {
-	this.tmp = this.read(this.addr) - (this.A & this.X);
-	this.fnzb(this.tmp);
-	this.X = (this.tmp & 0xFF);
-}
-
-CPU6502.prototype.sec = function() { this.C = 1; }
-CPU6502.prototype.sed = function() { this.D = 1; }
-CPU6502.prototype.sei = function() { this.I = 1; }
-
-CPU6502.prototype.shs = function() {
-	this.tmp = ((this.addr >> 8) + 1) & this.A & this.X;
-	this.write(this.addr, this.tmp & 0xFF);
-	this.S = (this.tmp & 0xFF);
-}
-
-CPU6502.prototype.shx = function() {
-	this.tmp = ((this.addr >> 8) + 1) & this.X;
-	this.write(this.addr, this.tmp & 0xFF);
-}
-
-CPU6502.prototype.shy = function() {
-	this.tmp = ((this.addr >> 8) + 1) & this.Y;
-	this.write(this.addr, this.tmp & 0xFF);
-}
-
-CPU6502.prototype.slo = function() {
-	this.tmp = this.read(this.addr) << 1;
-	this.tmp |= this.A;
-	this.fnzc(this.tmp);
-	this.A = this.tmp & 0xFF;
-}
-
-CPU6502.prototype.sre = function() {
-	var v = this.read(this.addr);
-	this.tmp = ((v & 1) << 8) | (v >> 1);
-	this.tmp ^= this.A;
-	this.fnzc(this.tmp);
-	this.A = this.tmp & 0xFF;
-}
-
-CPU6502.prototype.sta = function() {
-	this.write(this.addr, this.A);
-}
-
-CPU6502.prototype.stx = function() {
-	this.write(this.addr, this.X);
-}
-
-CPU6502.prototype.sty = function() {
-	this.write(this.addr, this.Y);
-}
-
-CPU6502.prototype.tax = function() {
-	this.X = this.A;
-	this.fnz(this.X);
-}
-
-CPU6502.prototype.tay = function() {
-	this.Y = this.A;
-	this.fnz(this.Y);
-}
-
-CPU6502.prototype.tsx = function() {
-	this.X = this.S;
-	this.fnz(this.X);
-}
-
-CPU6502.prototype.txa = function() {
-	this.A = this.X;
-	this.fnz(this.A);
-}
-
-CPU6502.prototype.txs = function() {
-	this.S = this.X;
-}
-
-CPU6502.prototype.tya = function() {
-	this.A = this.Y;
-	this.fnz(this.A);
-}
-
 ////////////////////////////////////////////////////////////////////////////////
 // Opcode table
 ////////////////////////////////////////////////////////////////////////////////
@@ -914,27 +276,4 @@ var CPU6502op = new Array();
 /*  INC abx */ CPU6502op[0xFE] = function(m) { m.abx(); m.inc(); m.rmw(); };
 /* *ISC abx */ CPU6502op[0xFF] = function(m) { m.abx(); m.isc(); m.rmw(); };
 
-////////////////////////////////////////////////////////////////////////////////
-// CPU instantiation
-////////////////////////////////////////////////////////////////////////////////
-
-var cpu = new CPU6502();
-
-cpu.read = function(addr) {
-	return mem[addr & 0xFFFF];
-}
-cpu.write = function(addr, value) {
-	mem[addr & 0xFFFF] = value;
-}
-
-cpu.reset();
-
-var start = Date.now();
-cpu.cycles = 0;
-
-while(cpu.cycles < 1000000) {
-	cpu.step();
-	//cpu.log();
-}
-console.log(cpu.cycles);
-console.log(Date.now() - start);
+export default CPU6502op;
