+++ webpack.dev.js
@@ -1,0 +1,10 @@
+const merge = require('webpack-merge');
+const common = require('./webpack.common.js');
+
+module.exports = merge(common, {
+    devtool: 'inline-source-map',
+    output:{
+        filename: '6502.dev.js'
+    }
+});
+
